Add working settings_reset firmware
Per @honorless:

```
This settings reset file was built from commit 2adaa00d, which predates the PR 2072 merge but post-dates the Zephyr 3.5 update. It also has USB logging enabled, so if you try flashing it please start logging when you're in bootloader.
If there is some sort of problem with the memory, hopefully some error about that would be shown around the time the firmware normally clears all existing BLE bond information (which is very early)
```

https://discord.com/channels/719497620560543766/1219493309429321728/1219639834281513100